# feat(response): improve compression with brotli and zstd support