--- a/upload/system/config/default.php
+++ b/upload/system/config/default.php
@@ -1,16 +1,16 @@
 <?php
 // Site
-$_['site_url']             = '';
+$_['site_url']                   = '';
 
 // Language
-$_['language_code']        = 'en-gb';
+$_['language_code']              = 'en-gb';
 
 // Date
-$_['date_timezone']        = 'UTC';
+$_['date_timezone']              = 'UTC';
 
 // Database
-$_['db_autostart']         = false;
-$_['db_option']            = [
+$_['db_autostart']               = false;
+$_['db_option']                  = [
 	'engine'   => 'mysqli', // mysqli, pdo or pgsql
 	'hostname' => 'localhost',
 	'username' => 'root',
@@ -23,50 +23,51 @@ $_['db_option']            = [
 ];
 
 // Mail
-$_['mail_engine']          = 'mail'; // mail or smtp
-$_['mail_from']            = ''; // Your E-Mail
-$_['mail_sender']          = ''; // Your name or company name
-$_['mail_reply_to']        = ''; // Reply to E-Mail
-$_['mail_smtp_hostname']   = '';
-$_['mail_smtp_username']   = '';
-$_['mail_smtp_password']   = '';
-$_['mail_smtp_port']       = 25;
-$_['mail_smtp_timeout']    = 5;
-$_['mail_verp']            = false;
-$_['mail_parameter']       = '';
+$_['mail_engine']                = 'mail'; // mail or smtp
+$_['mail_from']                  = ''; // Your E-Mail
+$_['mail_sender']                = ''; // Your name or company name
+$_['mail_reply_to']              = ''; // Reply to E-Mail
+$_['mail_smtp_hostname']         = '';
+$_['mail_smtp_username']         = '';
+$_['mail_smtp_password']         = '';
+$_['mail_smtp_port']             = 25;
+$_['mail_smtp_timeout']          = 5;
+$_['mail_verp']                  = false;
+$_['mail_parameter']             = '';
 
 // Cache
-$_['cache_engine']         = 'file'; // apc, file, mem, memcached or redis
-$_['cache_expire']         = 3600;
+$_['cache_engine']               = 'file'; // apc, file, mem, memcached or redis
+$_['cache_expire']               = 3600;
 
 // Session
-$_['session_autostart']    = false;
-$_['session_engine']       = 'file'; // db or file
-$_['session_name']         = 'OCSESSID';
-$_['session_domain']       = '';
-$_['session_path']         = !empty($_SERVER['PHP_SELF']) ? rtrim(dirname($_SERVER['PHP_SELF']), '/') . '/' : '/';
-$_['session_expire']       = 86400;
-$_['session_probability']  = 1;
-$_['session_divisor']      = 5;
-$_['session_samesite']     = 'Strict';
+$_['session_autostart']          = false;
+$_['session_engine']             = 'file'; // db or file
+$_['session_name']               = 'OCSESSID';
+$_['session_domain']             = '';
+$_['session_path']               = !empty($_SERVER['PHP_SELF']) ? rtrim(dirname($_SERVER['PHP_SELF']), '/') . '/' : '/';
+$_['session_expire']             = 86400;
+$_['session_probability']        = 1;
+$_['session_divisor']            = 5;
+$_['session_samesite']           = 'Strict';
 
 // Template
-$_['template_engine']      = 'twig';
-$_['template_extension']   = '.twig';
+$_['template_engine']            = 'twig';
+$_['template_extension']         = '.twig';
 
 // Error
-$_['error_display']        = true; // You need to change this to false on a live site.
-$_['error_log']            = true;
-$_['error_debug']          = false;
-$_['error_filename']       = 'error.log';
-$_['error_page']           = 'error.html';
+$_['error_display']              = true; // You need to change this to false on a live site.
+$_['error_log']                  = true;
+$_['error_debug']                = false;
+$_['error_filename']             = 'error.log';
+$_['error_page']                 = 'error.html';
 
 // Response
-$_['response_header']      = ['Content-Type: text/html; charset=utf-8'];
-$_['response_compression'] = 0;
+$_['response_header']            = ['Content-Type: text/html; charset=utf-8'];
+$_['response_compression']       = 0;
+$_['response_min_compress_size'] = 1024; // Minimum size for compression (bytes)
 
 // Actions
-$_['action_default']       = 'common/home';
-$_['action_error']         = 'error/not_found';
-$_['action_pre_action']    = [];
-$_['action_event']         = [];
+$_['action_default']             = 'common/home';
+$_['action_error']               = 'error/not_found';
+$_['action_pre_action']          = [];
+$_['action_event']               = [];