### Overview
Add support for modern compression algorithms with proper Accept-Encoding header parsing and automatic algorithm selection.

### Changes
- ✅ Add support for zstd and brotli compression
- ✅ Implement proper Accept-Encoding parsing with q-values and wildcard support
- ✅ Add priority-based algorithm selection (zstd > brotli > gzip > deflate)
- ✅ Add minimum compression size threshold to avoid CPU overhead on small files
- ✅ Add removeHeader() method for proper Content-Length management
- ✅ Add Vary: Accept-Encoding header for CDN/proxy compatibility
- ✅ Remove deprecated x-gzip support

### Test Results

#### Environment
- **Hardware**: MacBook Air 13-inch, M3 (8-core CPU: 4 performance + 4 efficiency; 8-core GPU), 16GB Unified Memory, 250GB SSD (2024)
- **Browser**: Firefox 140.0.4 (aarch64)
- **Setup**: OpenCart Docker Compose with `zstd` and `brotli` extensions
- **Test Page**: HTML home page (32.62 kB original size)

#### Compression Performance by Level

| Level | zstd (kB) | Brotli (kB) | Gzip (kB) | Best Algorithm |
|-------|-----------|-------------|-----------|----------------|
| 1     | 5.32      | 5.72        | 5.85      | **zstd**       |
| 2     | 5.37      | 5.19        | 5.70      | **brotli**     |
| 3     | 5.26      | 5.06        | 5.54      | **brotli**     |
| 4     | 5.27      | 4.76        | 5.21      | **brotli**     |
| 5     | 4.96      | 4.48        | 5.02      | **brotli**     |
| 6     | 4.90      | 4.46        | 4.93      | **brotli**     |
| 7     | 4.88      | 4.42        | 4.89      | **brotli**     |
| 8     | 4.85      | 4.41        | 4.88      | **brotli**     |
| 9     | 4.82      | 4.41        | 4.88      | **brotli**     |

#### Compression Ratios (Level 6 - Recommended)

| Algorithm | Original Size | Compressed Size | Compression Ratio | Bandwidth Saved |
|-----------|---------------|-----------------|-------------------|-----------------|
| **None**  | 32.62 kB      | 32.62 kB        | 0%                | 0%              |
| **Gzip**  | 32.62 kB      | 4.93 kB         | 84.9%             | 84.9%           |
| **Brotli**| 32.62 kB      | 4.46 kB         | 86.3%             | 86.3%           |
| **Zstd**  | 32.62 kB      | 4.90 kB         | 85.0%             | 85.0%           |

### How to Reproduce

#### Setup Requirements
1. Install PHP extensions: `zstd` and `brotli`
2. In `upload/system/config/default.php`:
   - Find `$_['response_compression'] = 0;` and change value to 1-9
   - Add `$_['response_min_compress_size'] = 1024;` (optional, for minimum compression size)

#### Testing Method
1. Set compression level in config: `$_['response_compression'] = 6;`
2. Temporarily disable algorithms in `getBestEncoding()` method to force specific compression
3. Test each compression level individually  
4. Measure in Firefox Dev Tools → Network tab → Size vs Transferred

![image](https://github.com/user-attachments/assets/e970786c-5a14-407d-a22d-9f47701eba0e)

### Performance Testing Results (Apache Benchmark)

**Test Configuration:**
- **Requests**: 1000 total, 10 concurrent
- **Hardware**: MacBook Air M3 (8-core CPU, 16GB RAM)  
- **Compression Level**: 6 (recommended balanced setting)
- **Test Page**: OpenCart home page (32.6 KB uncompressed)

| Algorithm | RPS | Document Size | Compression Ratio | Performance Impact | Comments |
|-----------|-----|---------------|-------------------|-------------------|----------|
| None      | 40.06 | 32618 bytes | 0% | Baseline | No compression |
| **Zstd**  | 30.78 | 4119 bytes | 87.4% | -23% RPS | Best balance |
| **Brotli**| 24.49 | 3675 bytes | 88.7% | -39% RPS | Best compression |
| **Gzip**  | 21.36 | 4146 bytes | 87.3% | -47% RPS | Legacy fallback |

**Performance Analysis:**
- **Zstd (Level 6)**: Optimal choice for production - excellent compression with minimal CPU overhead
- **Brotli (Level 6)**: Best compression ratio but higher CPU cost
- **Gzip (Level 6)**: Surprisingly CPU-intensive, validates priority placement below zstd/brotli
- **Bandwidth Savings**: 87-89% reduction in data transfer across all algorithms

#### How to Reproduce Apache Benchmark

```bash
docker run --rm --network host httpd:alpine ab -n 1000 -c 10 http://localhost/

echo "=== GZIP ===" && docker run --rm --network host httpd:alpine ab -n 1000 -c 10 -H "Accept-Encoding: gzip" http://localhost/ | grep -E "(Requests per second|Document Length|Transfer rate)"

echo "=== BROTLI ===" && docker run --rm --network host httpd:alpine ab -n 1000 -c 10 -H "Accept-Encoding: br" http://localhost/ | grep -E "(Requests per second|Document Length|Transfer rate)"

echo "=== ZSTD ===" && docker run --rm --network host httpd:alpine ab -n 1000 -c 10 -H "Accept-Encoding: zstd" http://localhost/ | grep -E "(Requests per second|Document Length|Transfer rate)"
```

### Conclusion

While **brotli showed better compression ratios** in these tests, **zstd is considered a more advanced algorithm** with better performance characteristics in production environments.

> [!NOTE]
> #### Important Notes
> - **Testing limitations**: Measurements were performed without server load and CPU usage was not measured
> - **Server-level compression recommended**: While PHP/OpenCart compression is useful and convenient, **server-level compression (nginx, Apache) is recommended** for production environments for better performance
> - **Algorithm priority**: Despite test results, zstd maintains highest priority due to its superior speed/compression balance under load
> - **Automatic selection**: The implementation automatically selects the best available algorithm based on browser support and server capabilities 

#### Production Recommendations
1. **Primary**: Use server-level compression (nginx gzip/brotli modules)
2. **Fallback**: Enable this PHP compression for dynamic content not handled by server
3. **Monitoring**: Monitor CPU usage in production when enabling higher compression levels

### References
- [RFC 9110 Accept-Encoding](https://www.rfc-editor.org/rfc/rfc9110.html#name-accept-encoding) - HTTP Accept-Encoding specification
- [Zstandard Compression Algorithm](https://facebook.github.io/zstd/) - Facebook's official documentation
- [Brotli Compression Format](https://github.com/google/brotli) - Google's official repository  
- [HTTP Compression - MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Compression) - Comprehensive overview
- [Can I Use brotli](https://caniuse.com/brotli)
- [Can I Use zstd](https://caniuse.com/zstd)

### Future Improvements

If this PR is accepted, there are several areas for potential enhancement:

#### Compression Level Optimization

Currently, compression levels are limited to 0-9 for compatibility. However, different algorithms support different ranges:

| Algorithm | Current Range | Native Range | Recommendation |
|-----------|---------------|--------------|----------------|
| **Gzip**  | 0-9 ✅       | 0-9          | Optimal as-is  |
| **Deflate** | 0-9 ✅     | 0-9          | Optimal as-is  |
| **Brotli** | 0-9 ⚠️      | 0-11         | Consider extending to 0-11 |
| **Zstd**  | 0-9 ⚠️       | -7 to 22     | Consider extending to 0-22 |

**References:**
- [Zstd compression levels documentation](https://facebook.github.io/zstd/zstd_manual.html) - supports levels 1-22
- [Brotli quality levels](https://github.com/google/brotli/issues/725) - supports levels 0-11
- [Cloudflare Brotli implementation](https://blog.cloudflare.com/this-is-brotli-from-origin/) - uses level 11 in production

#### Proposed Enhancement

```php
private function getOptimalLevel(string $encoding, int $requested_level): int {
    return match($encoding) {
        'zstd' => min($requested_level, 22),
        'br' => min($requested_level, 11), 
        'gzip', 'deflate' => min($requested_level, 9),
        default => $requested_level
    };
}
```

These improvements would be suitable for future PRs to maintain focused, reviewable changes.