@danielkerr 2025-07-13
Not a chance. Its not required as the store admins will rarely use the zip stuff.